backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
openai==1.59.9
openpyxl==3.1.5
python-dotenv==1.0.1
python-multipart==0.0.20
